refactor: weight 자료형 double로 변경

+++ src/main/java/PetBridge/adoptionPost/dto/AdoptionPostCreateDTO.java
@@ -18,7 +18,7 @@ public record AdoptionPostCreateDTO (
     String subTitle,
 
     @PositiveOrZero(message = "몸무게는 0 이상이어야 합니다.")
-    Long weight,
+    Double weight,
 
     @PositiveOrZero(message = "나이는 0 이상이어야 합니다.")
     Long age,
